(svn r13685) [0.6] -Backport from trunk: - Fix: In the case that elrails and 'realistic' acceleration are disabled all electrified engines would have no power on load, until the vehicle got turned around, loaded or got into a depot [FS#2102] (r13681) - Fix: Saving TTD imported games in recession failed due to wrong (and unneeded) type conversions in the saveload code [FS#2131] (r13679) - Fix: Inactive companies from old (TTD) saves could be marked active in some cases, which then loads garbage in their statistics and such [FS#2126] (r13676) - Fix: Desync when building electrified trains on a dedicated server that was started with electrification disabled [FS#2122] (r13673) - Fix: Clear the memory for the new AI during the loading of a savegame so it does not try to execute commands generated in a different savegame, which could be resulting in the AI trying to give orders to stations that do not exist (r13505) - Fix: Loading of very old OpenTTD savegames was broken (r13373)

+++ src/economy_type.h
@@ -15,7 +15,7 @@ struct Economy {
 	Money max_loan;         ///< Maximum possible loan
 	Money max_loan_unround; ///< Economy fluctuation status
 	uint16 max_loan_unround_fract; ///< Fraction of the unrounded max loan
-	int fluct;
+	int16 fluct;
 	byte interest_rate;     ///< Interest
 	byte infl_amount;       ///< inflation amount
 	byte infl_amount_pr;    ///< "floating" portion of inflation
